move resolver to core
